refactor(server): dissolve the platform folder into hmr, app and terminal
The platform/ directory is gone: its hot-update machinery joins hmr/, its
terminal tree joins the existing terminal/, and business.ts's two builders go
back into app.ts beside the runtime shell they were split from. hmr/ is left
holding the mechanism and the resource definitions the platform claims.

buildBusinessDeps and createBusinessApp keep their role — hmr/platform.ts's
create() still calls them per App, so a hot push still replaces the whole
business surface. That makes app.ts <-> hmr/platform.ts a module cycle, which
is safe here: both are hoisted function declarations, called only at App
creation, long after either module body has been evaluated.
