+++ packages/server/src/hmr/capabilities.ts
