+++ packages/server/src/terminal/input-mode.ts
