+++ packages/server/src/terminal/output-coalescer.ts
