+++ packages/server/src/terminal/shell.ts
