+++ packages/server/src/terminal/snapshot.ts
