+++ packages/server/src/terminal/stream.ts
