+++ packages/server/tsup.config.ts
@@ -6,7 +6,7 @@ export default defineConfig({
   // reset-admin-password are side-effect-free for CLI pre-checks and the offline
   // admin-password rescue); "./hmr/manifest" is the host-less harness.json reader the
   // CLI's thin loader resolves the current cli bundle through. There is deliberately no
-  // "./platform" subpath: src/platform/entry.ts (this package's platform artifact
+  // "./platform" subpath: src/hmr/entry.ts (this package's platform artifact
   // compile target — see its own module doc) is compiled straight from source by
   // esbuild in scripts/deploy.mjs, never through this package's published dist,
   // so it has no reason to be a tsup entry or a subpath export.
